+++ src/http/server.cpp
@@ -295,9 +295,9 @@ namespace http {
 
         if (request.getVersion() == "HTTP/1.1") {
             pResponse = version::handler_1_1::genResponse(request);
-        } else if (request.getVersion() == "HTTP/1.0") {
+        } else if (request.getVersion() == "HTTP/1.0" && conf::ENABLE_LEGACY_HTTP) {
             pResponse = version::handler_1_0::genResponse(request);
-        } else if (request.getVersion() == "HTTP/0.9") {
+        } else if (request.getVersion() == "HTTP/0.9" && conf::ENABLE_LEGACY_HTTP) {
             pResponse = version::handler_0_9::genResponse(request);
         } else {
             pResponse = new Response("HTTP/1.1"); // Default version