+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## v0.11.0
+- Now checks for a new version at startup (#91)
+    - Can be toggled off in mercury.conf
+    - Can now toggle startup welcome banner
+- Added support for toggling legacy HTTP version support (HTTP/0.9, HTTP/1.0) (#79)
+- Now makes immediate log file directory if missing (#5)
+- Updated default public HTML (#92)
+
 ## v0.10.0
 - Modularized request handlers (#39)
     - Added foundation for implementing future HTTP versions > HTTP/1.1